Struct ic_state_layout::state_layout::StateLayout

pub struct StateLayout { /* fields omitted */ }

StateLayout provides convenience functions to construct correct paths to individual components of the replicated execution state. It also utilizes filesystem specific checkpoint managers to create and manage checkpoints.

Following layout is overlayed on top of the filesystem specific checkpoint manager's internal layout. "checkpoints" directory is not owned by StateLayout and is internal to checkpoint manager

<root>
├── states_metadata.pbuf
│
│── tip
│   ├── system_metadata.pbuf
│   ├── subnet_queues.pbuf
│   └── canister_states
│       └── <hex(canister_id)>
│           ├── queues.pbuf
│           ├── vmemory_0.bin
│           ├── canister.pbuf
│           ├── stable_memory.(pbuf|bin)
│           └── software.wasm
│
├── [checkpoints] {owned and varies by checkpoint manager}
│   └──<hex(round)>
│      ├── system_metadata.pbuf
│      ├── subnet_queues.pbuf
│      └── canister_states
│          └── <hex(canister_id)>
│              ├── queues.pbuf
│              ├── vmemory_0.bin
│              ├── canister.pbuf
│              ├── stable_memory.(pbuf|bin)
│              └── software.wasm
│
└── tmp

Needs to be pub for criterion performance regression tests.

Implementations

impl StateLayout

pub fn new(log: ReplicaLogger, root: PathBuf) -> Self

Needs to be pub for criterion performance regression tests.

pub fn raw_path(&self) -> &Path

Returns the the raw root path for state

pub fn tmp(&self) -> Result<PathBuf, LayoutError>

Returns the path to the temporary directory. This directory is cleaned during restart of a node.

pub fn tip(&self) -> Result<CheckpointLayout<RwPolicy>, LayoutError>

Returns a layout object representing tip state in "tip" directory. During round execution this directory may contain inconsistent state. During full checkpointing this directory contains full state and is converted to a checkpoint. This directory is cleaned during restart of a node and reset to last full checkpoint.

pub fn states_metadata(&self) -> PathBuf

Returns the path to the serialized states metadata.

pub fn state_sync_scratchpad(
&self,
height: Height
) -> Result<PathBuf, LayoutError>

Returns scratchpad used during statesync

pub fn cleanup_tip(&self) -> Result<(), LayoutError>

pub fn tip_to_checkpoint(
&self,
tip: CheckpointLayout<RwPolicy>,
height: Height
) -> Result<CheckpointLayout<ReadOnly>, LayoutError>

Creates a checkpoint from the "tip" state returning layout object of the newly created checkpoint

pub fn scratchpad_to_checkpoint(
&self,
layout: CheckpointLayout<RwPolicy>,
height: Height
) -> Result<CheckpointLayout<ReadOnly>, LayoutError>

pub fn checkpoint_to_scratchpad(
&self,
height: Height
) -> Result<CheckpointLayout<RwPolicy>, LayoutError>

pub fn reset_tip_to(&self, height: Height) -> Result<(), LayoutError>

Resets "tip" to a checkpoint identified by height.

pub fn checkpoint(
&self,
height: Height
) -> Result<CheckpointLayout<ReadOnly>, LayoutError>

Returns the layout of the checkpoint with the given height (if there is one).

pub fn checkpoint_heights(&self) -> Result<Vec<Height>, LayoutError>

Returns a sorted list of Heights for which a checkpoint is available.

pub fn diverged_checkpoint_heights(&self) -> Result<Vec<Height>, LayoutError>

Returns a sorted list of Heights of checkpoints that were marked as diverged.

pub fn backup_heights(&self) -> Result<Vec<Height>, LayoutError>

Returns a sorted list of heights of checkpoints that were "backed up" for future inspection because they corresponded to diverged states.

pub fn diverged_checkpoint_path(&self, h: Height) -> PathBuf

Returns a path to a diverged checkpoint given its height.

If there is no diverged checkpoint with the specified height, the returned path doesn't exist on the filesystem.

Precondition: h ∈ self.diverged_checkpoint_heights()

pub fn backup_checkpoint_path(&self, h: Height) -> PathBuf

Returns a path to a backed up state given its height.

Precondition: h ∈ self.backup_heights()

pub fn remove_checkpoint(&self, height: Height) -> Result<(), LayoutError>

Removes a checkpoint for a given height if it exists.

Postcondition: height ∉ self.checkpoint_heights()

pub fn mark_checkpoint_diverged(
&self,
height: Height
) -> Result<(), LayoutError>

Marks the checkpoint with the specified height as diverged.

Precondition: height ∈ self.checkpoint_heights()

Postcondition: height ∈ self.diverged_checkpoint_heights() ∧ height ∉ self.checkpoint_heights()

pub fn remove_diverged_checkpoint(
&self,
height: Height
) -> Result<(), LayoutError>

Removes a diverged checkpoint given its height.

Precondition: h ∈ self.diverged_checkpoint_heights()

pub fn backup_checkpoint(&self, height: Height) -> Result<(), LayoutError>

Creates a copy of the checkpoint with the specified height and places it into a location that is not affected by normal state removal requests.

This is mostly useful for pinning good fetched states with the same height as the locally computed diverged ones. This makes it possible to check the difference between these two states and debug the non-determinism even if the whole subnet moved forward.

pub fn remove_backup(&self, height: Height) -> Result<(), LayoutError>

Removes a backed up state given its height.

Precondition: h ∈ self.backup_heights()

pub fn tip_path(&self) -> PathBuf