Trait ic_state_layout::state_layout::AccessPolicy

pub trait AccessPolicy {
    fn check_dir(_p: &Path) -> Result<(), LayoutError> { ... }
}

Provided methods

fn check_dir(_p: &Path) -> Result<(), LayoutError>

check_dir specifies what to do the first time we enter a directory while reading/writing a checkpoint.

The default behavior is to do nothing. This is suitable for the ReadOnly mode because if the directory doesn't exist, we'll fail anyway when we try to read files from it.

Loading content...