Trait ic_state_layout::state_layout::CheckpointManager

pub trait CheckpointManager: Send + Sync {
    fn raw_path(&self) -> &Path;

    fn tip_to_checkpoint(&self, tip: &Path, name: &str) -> Result<PathBuf>;

    fn remove_checkpoint(&self, name: &str) -> Result<()>;

    fn scratchpad_to_checkpoint(
        &self, 
        path: &Path, 
        name: &str
    ) -> Result<PathBuf>;

    fn checkpoint_to_scratchpad(&self, name: &str, path: &Path) -> Result<()>;

    fn get_checkpoint_path(&self, name: &str) -> PathBuf;

    fn mark_checkpoint_diverged(&self, name: &str) -> Result<()>;

    fn backup_checkpoint(&self, name: &str) -> Result<()>;

    fn list_diverged_checkpoints(&self) -> Result<Vec<String>>;

    fn list_backups(&self) -> Result<Vec<String>>;

    fn get_diverged_checkpoint_path(&self, name: &str) -> PathBuf;

    fn get_backup_path(&self, name: &str) -> PathBuf;

    fn remove_diverged_checkpoint(&self, name: &str) -> Result<()>;

    fn remove_backup(&self, name: &str) -> Result<()>;

    fn list_checkpoints(&self) -> Result<Vec<String>>;

    fn reset_tip_to(&self, tip: &PathBuf, name: &str) -> Result<()>;
}

Required methods

fn raw_path(&self) -> &Path

Returns the base directory path managed by checkpoint manager.

fn tip_to_checkpoint(&self, tip: &Path, name: &str) -> Result<PathBuf>

Atomically creates a checkpoint from the "tip" directory. "name" identifies the name of the checkpoint to be created and it need not be a directory path.

fn remove_checkpoint(&self, name: &str) -> Result<()>

Removes a checkpoint identified by "name".

fn scratchpad_to_checkpoint(&self, path: &Path, name: &str) -> Result<PathBuf>

Promotes the scratchpad at "path" to a checkpoint identified by "name".

fn checkpoint_to_scratchpad(&self, name: &str, path: &Path) -> Result<()>

Creates a mutable copy of a checkpoint in a the specified path.

fn get_checkpoint_path(&self, name: &str) -> PathBuf

Gets the path for a checkpoint identified by "name".

fn mark_checkpoint_diverged(&self, name: &str) -> Result<()>

Marks checkpoint identified by "name" as diverged.

Post-condition: name ∉ self.list_checkpoints()

fn backup_checkpoint(&self, name: &str) -> Result<()>

Creates a backup of the checkpoint identified by "name".

fn list_diverged_checkpoints(&self) -> Result<Vec<String>>

Returns the list of names of diverged checkpoints.

fn list_backups(&self) -> Result<Vec<String>>

Returns the list of names of checkpoint backups.

fn get_diverged_checkpoint_path(&self, name: &str) -> PathBuf

Returns a path to diverged checkpoint with the specified name.

fn get_backup_path(&self, name: &str) -> PathBuf

Returns a path to a backup the specified name.

fn remove_diverged_checkpoint(&self, name: &str) -> Result<()>

Removes the diverged checkpoint with the specified name.

fn remove_backup(&self, name: &str) -> Result<()>

Removes the backup with the specified name.

fn list_checkpoints(&self) -> Result<Vec<String>>

List names of all the existing checkpoints.

fn reset_tip_to(&self, tip: &PathBuf, name: &str) -> Result<()>

Atomically resets containts of "tip" directory to that of checkpoint

Loading content...